core: carry each Clip's Overlays on every query the export address reads
The Export Hash is about to fold Overlays in, and it is computed from Clip
rows handed to it by these six queries. A query that left Overlays out would
report a Video as having none, so its address would go on claiming that a
Definition Card edit never happened and no Publish would re-render it.

One shared `overlayExportRelation` fragment rather than six inline copies, so
the queries cannot drift apart.

db-video-operations was already over the 5500-token file budget and could not
take another line, so the next/previous-Video walks — which touch no table at
all, only the Course navigation tree — move to their own module unchanged.
